+++ chains/ethereum/index.ts
@@ -1,4 +1,4 @@
 export type { Chain, ChainProviderFn, Connector, Provider, WebSocketProvider } from '@wagmi/core'
 export { Web3ModalEthereum } from './src/api'
-export { modalChains, modalConnectors, modalProviders } from './src/utils/wagmiTools'
+export { modalConnectors, walletConnectProvider } from './src/wagmiTools'
 export * from './types/apiTypes'
