+++ examples/html/package.json
@@ -11,7 +11,8 @@
   "dependencies": {
     "@web3modal/core": "2.0.0-a38e60a",
     "@web3modal/ethereum": "2.0.0-a38e60a",
-    "@web3modal/ui": "2.0.0-a38e60a"
+    "@web3modal/ui": "2.0.0-a38e60a",
+    "@wagmi/core": "0.7.2"
   },
   "devDependencies": {
     "vite": "3.2.3"
