+++ examples/html/src/main.js
@@ -1,25 +1,28 @@
+import { chain, configureChains, createClient } from '@wagmi/core'
 import { ClientCtrl, ConfigCtrl } from '@web3modal/core'
-import { modalChains, modalProviders } from '@web3modal/ethereum'
+import { modalConnectors, walletConnectProvider, Web3ModalEthereum } from '@web3modal/ethereum'
 import '@web3modal/ui'
 import './actions.js'
 import './events.js'
 
 // Define constants
-const PROJECT_ID = '8e6b5ffdcbc9794bf9f4a1952578365b'
+const projectId = '8e6b5ffdcbc9794bf9f4a1952578365b'
+const chains = [chain.mainnet]
 
-const clientConfig = {
-  projectId: PROJECT_ID,
-  theme: 'dark',
-  accentColor: 'default'
-}
+const { provider } = configureChains(chains, [walletConnectProvider({ projectId })])
 
-const ethereumConfig = {
-  appName: 'web3Modal',
+const wagmiClient = createClient({
   autoConnect: true,
-  chains: [modalChains.mainnet],
-  providers: [modalProviders.walletConnectProvider({ projectId: PROJECT_ID })]
-}
+  connectors: modalConnectors({ appName: 'web3Moda', chains }),
+  provider
+})
+
+const ethereumClient = Web3ModalEthereum.create(wagmiClient, chains)
 
 // Set up core and ethereum clients
-ConfigCtrl.setConfig(clientConfig)
-ClientCtrl.setEthereumClient(ethereumConfig)
+ConfigCtrl.setConfig({
+  projectId,
+  theme: 'dark',
+  accentColor: 'default'
+})
+ClientCtrl.setEthereumClient(ethereumClient, chains)
