+++ examples/react/src/pages/_app.tsx
@@ -1,12 +1,7 @@
-import {
-  modalChains,
-  modalConnectors,
-  modalProviders,
-  Web3ModalEthereum
-} from '@web3modal/ethereum'
+import { modalConnectors, walletConnectProvider, Web3ModalEthereum } from '@web3modal/ethereum'
 import { Web3Modal } from '@web3modal/react'
 import type { AppProps } from 'next/app'
-import { configureChains, createClient, WagmiConfig } from 'wagmi'
+import { chain, configureChains, createClient, WagmiConfig } from 'wagmi'
 import '../styles.css'
 
 // Get projectID at https://cloud.walletconnect.com
@@ -14,18 +9,9 @@ if (!process.env.NEXT_PUBLIC_PROJECT_ID)
   throw new Error('You need to provide NEXT_PUBLIC_PROJECT_ID env variable')
 
 const projectId = process.env.NEXT_PUBLIC_PROJECT_ID
+const chains = [chain.mainnet, chain.polygon, chain.optimism, chain.arbitrum]
 
-const chains = [
-  modalChains.mainnet,
-  modalChains.polygon,
-  modalChains.avalanche,
-  modalChains.fantom,
-  modalChains.optimism,
-  modalChains.arbitrum,
-  modalChains.binanceSmartChain
-]
-
-const { provider } = configureChains(chains, [modalProviders.walletConnectProvider({ projectId })])
+const { provider } = configureChains(chains, [walletConnectProvider({ projectId })])
 
 const wagmiClient = createClient({
   autoConnect: true,
